🔧 chore(.firebaserc): add "mycure-prm" as a production target for "mc-v4-prod" project 🔧 chore(firebase-hosting-production.yml): set target to "production" for Firebase Hosting deployment 🔧 chore(firebase-hosting-pull-request.yml): set target to "production" for Firebase Hosting deployment in pull requests
The changes were made to configure the Firebase project to include "mycure-prm" as a production target for the "mc-v4-prod" project. This allows the application to be deployed to the "mycure-prm" hosting environment when deploying to production. The target configuration was also updated in the Firebase Hosting deployment workflows to ensure that the application is deployed to the correct target.