🔧 chore(firebase-hosting-pull-request.yml): update APP_LINK_CMS_APP environment variable to point to the correct URL 🔧 chore(firebase-hosting-staging.yml): update APP_LINK_CMS_APP environment variable to point to the correct URL 🔧 chore(quasar.config.js): refactor environment variable configuration to use dotenv package and display environment variables
The APP_LINK_CMS_APP environment variable in both the firebase-hosting-pull-request.yml and firebase-hosting-staging.yml files has been updated to point to the correct URL (https://twitter.com/mycuremd) instead of the previous URL (https://staging-web-cms.web.app). This change ensures that the correct link to the CMS app is used in the deployment.

In the quasar.config.js file, the configuration for environment variables has been refactored to use the dotenv package. The environment variables are now loaded using dotenv and displayed in the console for debugging purposes. This change improves the readability and maintainability of the configuration file.