🐛 fix(.env.example): update APP_LINK_TWITTER value to correct Twitter account 🚀 chore(firebase-hosting-production.yml): add step to create .env file with updated values 🚀 chore(firebase-hosting-pull-request.yml): add step to create .env file with updated values 🚀 chore(firebase-hosting-staging.yml): add step to create .env file with updated values
The APP_LINK_TWITTER value in the .env.example file was updated to the correct Twitter account. This change ensures that the correct Twitter account is linked in the application. Additionally, in the firebase-hosting-production.yml, firebase-hosting-pull-request.yml, and firebase-hosting-staging.yml files, steps were added to create the .env file with the updated values. This ensures that the correct environment variables are used during the build and deployment process.